scx_lavd: More accurate time slice calculation.
In the past, we unconditionally added +1 to the number of queued tasks to
avoid the divide-by-zero error. However, now that the time slice is shorter
than the slice_max_ns, the slice boost won't be considered. Let's stop
doing that; instead, calculate it precisely without blindly adding +1 to
the number of queued tasks.

Signed-off-by: Changwoo Min <changwoo@igalia.com>