ci: append nix paths to formatter
The formatter works fine in the CI for Nix files, but when I try to run
it locally it hangs forever. This is a bug in a specific version of Nix
that means it waits forever for stdin if given no paths (it should
format everything). Small quality of life improvement to add the Nix
paths so it doesn't hang and I can format things locally.

Test plan:
- CI
- Ran locally. Works now.