scx_chaos: add README.md
Add a README for chaos with a logo. The logo may need a little iteration
but I'm generally happy with it. The README is probably bad but the best
way to achieve correct things on the Internet is to wait for somebody to
correct it :)

Will do my best to keep this up to date.