scx_mitosis: dynamic cell creation/destruction
This commit adds dynamic cell creation and destruction. Cells are
defined by having a cpuset applied to a cgroup. The way we determine
this is through writes to cpuset.cpus (e.g. a tracepoint) or on
cgroup_init. There's a lot of concern about racy behavior here, so the
logic is anytime we see a change, we trigger a subsequent tick() to
actually evaluate which cells should exist. This enables the logic to be
serialized.

Signed-off-by: Dan Schatzberg <schatzberg.dan@gmail.com>