scx_lavd: Continue to use a CPU that has tasks on its per-CPU DSQ.
When there are tasks on a per-CPU DSQ, we should continue to use
that CPU until all the tasks are consumed. Otherwise, the tasks
on a per-CPU DSQ will be stalled. To this end, we check if there
is a task on a per-CPU DSQ in two places: 1) ops.dispatch() to
determine if this CPU should consume a task or not, and 2) core
compaction to determine if this CPU should be included in the
overflow set.

Signed-off-by: Changwoo Min <changwoo@igalia.com>