scx_lavd: Ignore an error from set_rlimit_infinity().
Setting rlimit for MEMLOCK to infinity would fail with -ENOPERM.
However, it does not necessarily mean catastrophic failure. Especially
when rlimits are already set to a large enough value, it is okay to
ignore the error.

So let's ignore the error and print the current rlimits. If the current
rlimits are insufficient, the program will fail anyway later in the
initialization, allocating BPF objects.

Signed-off-by: Changwoo Min <changwoo@igalia.com>