scx_mitosis: Clean up dispatch()
Currently, if iterator based peek fails in dispatch for both the cell
DSQ and the CPU DSQ, min_vtime_dsq may be used uninitialized. We are
guaranteed to have it set to valid DSQ if found = true, hence return
early if so (and document that the prev task if not NULL keeps running
due to the default behavior in absence of SCX_OPS_ENQ_LAST and does not
require explicit slice refresh).

Also, we don't need to attempt a move to local DSQ if min_vtime_dsq
matches cpu_dsq(cpu), hence guard with an additional check.

 [
   The behavior in kernel (ext.c:balance_one()) is as follows: if the slice
   is depleted when dispatch is called, it is continued (in absence of
   other tasks in local and global DSQ) iff the slice is refreshed, or the
   SCX_OPS_ENQ_LAST is missing (which retains default behavior before the
   addition of the flag). Of course, the task must not be dequeued from the
   rq before we hit dispatch (i.e., p->scx.flags & SCX_TASK_QUEUED == true).
 ]

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>