scx_layered: Sanity check HintEquals specification
Since the way matching logic works for HintEquals needs to support high
frequency detection, it cannot be expensive and complete. This implies
that if HintEquals matching clause is used in conjunction with another
matcher to create a matching condition, we would have to invoke the full
blown layer matching logic on every wake up to ensure correct layer
memberships. Since this would be prohibitively expensive, reject the use
of HintEquals in conjunction with other matchers in the same AND block.

It can be permitted to combine HintEquals in a disjunction and OR it
with other matchers, such that its presence alone is sufficient to
ensure we can decide which layer a task should be switched to. To this
end, add sanity checks that reject invalid configurations.

Signed-off-by: Kumar Kartikeya Dwivedi <memxor@gmail.com>