scx_raw_pmu: use include_dir to embed json
scx_raw_pmu needs a bunch of JSON files. Currently these are expected to
be distributed with the crate/scheduler that uses them, but this is
annoying.

Use the `include_dir!` macro to package all of these resources into
the crate's library, such that they can be used directly in other crates.
Add an abstraction `ResourceDir` which wraps both this `include_dir::Dir`
method and a regular Filesystem method for cases where the user wants to
override. It's not clear this is needed, but it's less of an API break.

Test plan:
- CI
- Enabled the existing test of the filesystem based method by extracting
  the included files into a tempdir.
- Added a new similar test that uses the included files from memory.